Update buildupdate scripts for new dependencies

+++ build/buildupdate.mono.sh
@@ -1,10 +1,25 @@
 #!/bin/bash
 # server=build.palaso.org
 # project=WeSay1.4-Linux
-# build=WeSay1.4-Precise64 Continuous
+# build=wesay1.4-precise64-continuous
 # root_dir=..
+# $Id: da666a7e5eb1d63b434514279cd14cacd26c730f $
 
 # *** Functions ***
+force=
+
+while getopts f opt; do
+	case $opt in
+	f)
+		force=1
+		;;
+
+	esac
+done
+
+shift $((OPTIND - 1))
+
+
 copy_auto() {
 	where_curl=$(type -P curl)
 	where_wget=$(type -P wget)
@@ -22,7 +37,7 @@ copy_auto() {
 
 copy_curl() {
 	echo "curl: $2 <= $1"
-	if [ -e "$2" ]
+	if [ -e "$2" ] && [ "$force" != "1" ]
 	then
 		curl -# -L -z $2 -o $2 $1
 	else
@@ -40,32 +55,32 @@ copy_wget() {
 }
 
 # *** Results ***
-# build: WeSay1.4-Precise64 Continuous (bt314)
+# build: wesay1.4-precise64-continuous (bt314)
 # project: WeSay1.4-Linux
 # URL: http://build.palaso.org/viewType.html?buildTypeId=bt314
 # VCS: https://github.com/sillsdev/wesay [develop]
 # dependencies:
-# [0] build: chorus-precise64-master Continuous (bt323)
+# [0] build: chorus-precise64-develop-continuous (bt337)
 #     project: Chorus
-#     URL: http://build.palaso.org/viewType.html?buildTypeId=bt323
+#     URL: http://build.palaso.org/viewType.html?buildTypeId=bt337
 #     clean: false
 #     revision: latest.lastSuccessful
-#     paths: {"Autofac.dll"=>"lib/common"}
-#     VCS: https://github.com/sillsdev/chorus.git [master]
-# [1] build: chorus-precise64-master Continuous (bt323)
-#     project: Chorus
-#     URL: http://build.palaso.org/viewType.html?buildTypeId=bt323
-#     clean: false
-#     revision: latest.lastSuccessful
-#     paths: {"Chorus.exe"=>"lib/Release", "LibChorus.dll"=>"lib/Release", "ChorusMerge.exe"=>"lib/Release", "LibChorus.TestUtilities.dll"=>"lib/Release", "Mercurial-i686.zip"=>"lib/Release"}
-#     VCS: https://github.com/sillsdev/chorus.git [master]
-# [2] build: L10NSharp Mono continuous (bt271)
+#     paths: {"release/**"=>"lib/Release", "debug/**"=>"lib/Debug", "Autofac.dll"=>"lib/common", "Mercurial-i686.zip"=>"lib/common", "Mercurial-x86_64.zip"=>"lib/common"}
+#     VCS: https://github.com/sillsdev/chorus.git [develop]
+# [1] build: L10NSharp Mono continuous (bt271)
 #     project: L10NSharp
 #     URL: http://build.palaso.org/viewType.html?buildTypeId=bt271
 #     clean: false
 #     revision: latest.lastSuccessful
-#     paths: {"L10NSharp.dll"=>"lib/common"}
+#     paths: {"L10NSharp.dll"=>"lib/common", "L10NSharp.dll.mdb"=>"lib/common"}
 #     VCS: https://bitbucket.org/hatton/l10nsharp [default]
+# [2] build: palaso-precise64-develop-continuous (bt334)
+#     project: libpalaso
+#     URL: http://build.palaso.org/viewType.html?buildTypeId=bt334
+#     clean: false
+#     revision: latest.lastSuccessful
+#     paths: {"release/**"=>"lib/Release", "debug/**"=>"lib/Debug"}
+#     VCS: https://github.com/sillsdev/libpalaso.git [develop]
 # [3] build: icucil-precise64-Continuous (bt281)
 #     project: Libraries
 #     URL: http://build.palaso.org/viewType.html?buildTypeId=bt281
@@ -73,34 +88,98 @@ copy_wget() {
 #     revision: latest.lastSuccessful
 #     paths: {"icu*.dll"=>"lib/Release", "icu*.config"=>"lib/Release"}
 #     VCS: https://github.com/sillsdev/icu-dotnet [master]
-# [4] build: palaso-precise64-master Continuous (bt322)
-#     project: libpalaso
-#     URL: http://build.palaso.org/viewType.html?buildTypeId=bt322
+# [4] build: icucil-precise64-Continuous (bt281)
+#     project: Libraries
+#     URL: http://build.palaso.org/viewType.html?buildTypeId=bt281
 #     clean: false
 #     revision: latest.lastSuccessful
-#     paths: {"Palaso.BuildTasks.dll"=>"lib/Release", "Palaso.dll"=>"lib/Release", "Palaso.TestUtilities.dll"=>"lib/Release", "Palaso.Tests.dll"=>"lib/Release", "Palaso.DictionaryServices.dll"=>"lib/Release", "Palaso.Media.dll"=>"lib/Release", "PalasoUIWindowsForms.dll"=>"lib/Release", "Palaso.Lift.dll"=>"lib/Release"}
-#     VCS: https://github.com/sillsdev/libpalaso.git [master]
+#     paths: {"icu*.dll"=>"lib/Debug", "icu*.config"=>"lib/Debug"}
+#     VCS: https://github.com/sillsdev/icu-dotnet [master]
 
 # make sure output directories exist
-mkdir -p ../lib/common
 mkdir -p ../lib/Release
+mkdir -p ../lib/Debug
+mkdir -p ../lib/common
 
 # download artifact dependencies
-copy_auto http://build.palaso.org/guestAuth/repository/download/bt323/latest.lastSuccessful/Autofac.dll ../lib/common/Autofac.dll
-copy_auto http://build.palaso.org/guestAuth/repository/download/bt323/latest.lastSuccessful/Chorus.exe ../lib/Release/Chorus.exe
-copy_auto http://build.palaso.org/guestAuth/repository/download/bt323/latest.lastSuccessful/LibChorus.dll ../lib/Release/LibChorus.dll
-copy_auto http://build.palaso.org/guestAuth/repository/download/bt323/latest.lastSuccessful/ChorusMerge.exe ../lib/Release/ChorusMerge.exe
-copy_auto http://build.palaso.org/guestAuth/repository/download/bt323/latest.lastSuccessful/LibChorus.TestUtilities.dll ../lib/Release/LibChorus.TestUtilities.dll
-copy_auto http://build.palaso.org/guestAuth/repository/download/bt323/latest.lastSuccessful/Mercurial-i686.zip ../lib/Release/Mercurial-i686.zip
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/release/Chorus.exe ../lib/Release/Chorus.exe
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/release/Chorus.exe.config ../lib/Release/Chorus.exe.config
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/release/Chorus.exe.mdb ../lib/Release/Chorus.exe.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/release/ChorusMerge.exe ../lib/Release/ChorusMerge.exe
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/release/ChorusMerge.exe.config ../lib/Release/ChorusMerge.exe.config
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/release/ChorusMerge.exe.mdb ../lib/Release/ChorusMerge.exe.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/release/Geckofx-Winforms-14.dll ../lib/Release/Geckofx-Winforms-14.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/release/LibChorus.TestUtilities.dll ../lib/Release/LibChorus.TestUtilities.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/release/LibChorus.TestUtilities.dll.mdb ../lib/Release/LibChorus.TestUtilities.dll.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/release/LibChorus.dll ../lib/Release/LibChorus.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/release/LibChorus.dll.mdb ../lib/Release/LibChorus.dll.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/release/NDesk.DBus.dll ../lib/Release/NDesk.DBus.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/release/NDesk.DBus.dll.config ../lib/Release/NDesk.DBus.dll.config
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/release/geckofix.so ../lib/Release/geckofix.so
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/release/geckofx-core-14.dll ../lib/Release/geckofx-core-14.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/release/geckofx-core-14.dll.config ../lib/Release/geckofx-core-14.dll.config
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/debug/Chorus.exe ../lib/Debug/Chorus.exe
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/debug/Chorus.exe.config ../lib/Debug/Chorus.exe.config
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/debug/Chorus.exe.mdb ../lib/Debug/Chorus.exe.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/debug/ChorusMerge.exe ../lib/Debug/ChorusMerge.exe
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/debug/ChorusMerge.exe.config ../lib/Debug/ChorusMerge.exe.config
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/debug/ChorusMerge.exe.mdb ../lib/Debug/ChorusMerge.exe.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/debug/Geckofx-Winforms-14.dll ../lib/Debug/Geckofx-Winforms-14.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/debug/LibChorus.TestUtilities.dll ../lib/Debug/LibChorus.TestUtilities.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/debug/LibChorus.TestUtilities.dll.mdb ../lib/Debug/LibChorus.TestUtilities.dll.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/debug/LibChorus.dll ../lib/Debug/LibChorus.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/debug/LibChorus.dll.mdb ../lib/Debug/LibChorus.dll.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/debug/NDesk.DBus.dll ../lib/Debug/NDesk.DBus.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/debug/NDesk.DBus.dll.config ../lib/Debug/NDesk.DBus.dll.config
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/debug/geckofix.so ../lib/Debug/geckofix.so
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/debug/geckofx-core-14.dll ../lib/Debug/geckofx-core-14.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/debug/geckofx-core-14.dll.config ../lib/Debug/geckofx-core-14.dll.config
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/Autofac.dll ../lib/common/Autofac.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/Mercurial-i686.zip ../lib/common/Mercurial-i686.zip
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt337/latest.lastSuccessful/Mercurial-x86_64.zip ../lib/common/Mercurial-x86_64.zip
 copy_auto http://build.palaso.org/guestAuth/repository/download/bt271/latest.lastSuccessful/L10NSharp.dll ../lib/common/L10NSharp.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt271/latest.lastSuccessful/L10NSharp.dll.mdb ../lib/common/L10NSharp.dll.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/release/Palaso.BuildTasks.dll ../lib/Release/Palaso.BuildTasks.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/release/Palaso.DictionaryServices.dll ../lib/Release/Palaso.DictionaryServices.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/release/Palaso.Lift.dll ../lib/Release/Palaso.Lift.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/release/Palaso.Lift.dll.mdb ../lib/Release/Palaso.Lift.dll.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/release/Palaso.Media.dll ../lib/Release/Palaso.Media.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/release/Palaso.Media.dll.config ../lib/Release/Palaso.Media.dll.config
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/release/Palaso.Media.dll.mdb ../lib/Release/Palaso.Media.dll.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/release/Palaso.TestUtilities.dll ../lib/Release/Palaso.TestUtilities.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/release/Palaso.TestUtilities.dll.mdb ../lib/Release/Palaso.TestUtilities.dll.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/release/Palaso.Tests.dll ../lib/Release/Palaso.Tests.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/release/Palaso.Tests.dll.mdb ../lib/Release/Palaso.Tests.dll.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/release/Palaso.dll ../lib/Release/Palaso.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/release/Palaso.dll.config ../lib/Release/Palaso.dll.config
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/release/Palaso.dll.mdb ../lib/Release/Palaso.dll.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/release/PalasoUIWindowsForms.dll ../lib/Release/PalasoUIWindowsForms.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/release/PalasoUIWindowsForms.dll.config ../lib/Release/PalasoUIWindowsForms.dll.config
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/release/PalasoUIWindowsForms.dll.mdb ../lib/Release/PalasoUIWindowsForms.dll.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/release/SIL.Archiving.dll ../lib/Release/SIL.Archiving.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/release/SIL.Archiving.dll.config ../lib/Release/SIL.Archiving.dll.config
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/release/SIL.Archiving.dll.mdb ../lib/Release/SIL.Archiving.dll.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/debug/Palaso.BuildTasks.dll ../lib/Debug/Palaso.BuildTasks.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/debug/Palaso.DictionaryServices.dll ../lib/Debug/Palaso.DictionaryServices.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/debug/Palaso.Lift.dll ../lib/Debug/Palaso.Lift.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/debug/Palaso.Lift.dll.mdb ../lib/Debug/Palaso.Lift.dll.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/debug/Palaso.Media.dll ../lib/Debug/Palaso.Media.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/debug/Palaso.Media.dll.config ../lib/Debug/Palaso.Media.dll.config
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/debug/Palaso.Media.dll.mdb ../lib/Debug/Palaso.Media.dll.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/debug/Palaso.TestUtilities.dll ../lib/Debug/Palaso.TestUtilities.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/debug/Palaso.TestUtilities.dll.mdb ../lib/Debug/Palaso.TestUtilities.dll.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/debug/Palaso.Tests.dll ../lib/Debug/Palaso.Tests.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/debug/Palaso.Tests.dll.mdb ../lib/Debug/Palaso.Tests.dll.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/debug/Palaso.dll ../lib/Debug/Palaso.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/debug/Palaso.dll.config ../lib/Debug/Palaso.dll.config
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/debug/Palaso.dll.mdb ../lib/Debug/Palaso.dll.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/debug/PalasoUIWindowsForms.dll ../lib/Debug/PalasoUIWindowsForms.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/debug/PalasoUIWindowsForms.dll.mdb ../lib/Debug/PalasoUIWindowsForms.dll.mdb
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/debug/SIL.Archiving.dll ../lib/Debug/SIL.Archiving.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/debug/SIL.Archiving.dll.config ../lib/Debug/SIL.Archiving.dll.config
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt334/latest.lastSuccessful/debug/SIL.Archiving.dll.mdb ../lib/Debug/SIL.Archiving.dll.mdb
 copy_auto http://build.palaso.org/guestAuth/repository/download/bt281/latest.lastSuccessful/icu.net.dll ../lib/Release/icu.net.dll
 copy_auto http://build.palaso.org/guestAuth/repository/download/bt281/latest.lastSuccessful/icu.net.dll.config ../lib/Release/icu.net.dll.config
-copy_auto http://build.palaso.org/guestAuth/repository/download/bt322/latest.lastSuccessful/Palaso.BuildTasks.dll ../lib/Release/Palaso.BuildTasks.dll
-copy_auto http://build.palaso.org/guestAuth/repository/download/bt322/latest.lastSuccessful/Palaso.dll ../lib/Release/Palaso.dll
-copy_auto http://build.palaso.org/guestAuth/repository/download/bt322/latest.lastSuccessful/Palaso.TestUtilities.dll ../lib/Release/Palaso.TestUtilities.dll
-copy_auto http://build.palaso.org/guestAuth/repository/download/bt322/latest.lastSuccessful/Palaso.Tests.dll ../lib/Release/Palaso.Tests.dll
-copy_auto http://build.palaso.org/guestAuth/repository/download/bt322/latest.lastSuccessful/Palaso.DictionaryServices.dll ../lib/Release/Palaso.DictionaryServices.dll
-copy_auto http://build.palaso.org/guestAuth/repository/download/bt322/latest.lastSuccessful/Palaso.Media.dll ../lib/Release/Palaso.Media.dll
-copy_auto http://build.palaso.org/guestAuth/repository/download/bt322/latest.lastSuccessful/PalasoUIWindowsForms.dll ../lib/Release/PalasoUIWindowsForms.dll
-copy_auto http://build.palaso.org/guestAuth/repository/download/bt322/latest.lastSuccessful/Palaso.Lift.dll ../lib/Release/Palaso.Lift.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt281/latest.lastSuccessful/icu.net.dll ../lib/Debug/icu.net.dll
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt281/latest.lastSuccessful/icu.net.dll.config ../lib/Debug/icu.net.dll.config
 # End of script
